Start cleaning up structure of code action code (#163875)
- Moves common types into `common`
- Move some command IDs into common
- Moves getCodeActions into `codeAction.ts` instead of `codeActionCommands`
