+++ src/vs/editor/contrib/codeAction/browser/codeAction.ts
@@ -6,19 +6,25 @@
 import { coalesce, equals, isNonEmptyArray } from 'vs/base/common/arrays';
 import { CancellationToken } from 'vs/base/common/cancellation';
 import { illegalArgument, isCancellationError, onUnexpectedExternalError } from 'vs/base/common/errors';
-import { Disposable, DisposableStore, IDisposable } from 'vs/base/common/lifecycle';
+import { Disposable, DisposableStore } from 'vs/base/common/lifecycle';
 import { URI } from 'vs/base/common/uri';
-import { TextModelCancellationTokenSource } from 'vs/editor/contrib/editorState/browser/editorState';
+import { ICodeEditor } from 'vs/editor/browser/editorBrowser';
+import { IBulkEditService } from 'vs/editor/browser/services/bulkEditService';
 import { Range } from 'vs/editor/common/core/range';
 import { Selection } from 'vs/editor/common/core/selection';
-import { ITextModel } from 'vs/editor/common/model';
-import * as languages from 'vs/editor/common/languages';
-import { IModelService } from 'vs/editor/common/services/model';
-import { CommandsRegistry } from 'vs/platform/commands/common/commands';
-import { IProgress, Progress } from 'vs/platform/progress/common/progress';
-import { CodeActionFilter, CodeActionKind, CodeActionTrigger, CodeActionTriggerSource, filtersAction, mayIncludeActionsOfKind } from './types';
 import { LanguageFeatureRegistry } from 'vs/editor/common/languageFeatureRegistry';
+import * as languages from 'vs/editor/common/languages';
+import { ITextModel } from 'vs/editor/common/model';
 import { ILanguageFeaturesService } from 'vs/editor/common/services/languageFeatures';
+import { IModelService } from 'vs/editor/common/services/model';
+import { TextModelCancellationTokenSource } from 'vs/editor/contrib/editorState/browser/editorState';
+import * as nls from 'vs/nls';
+import { CommandsRegistry, ICommandService } from 'vs/platform/commands/common/commands';
+import { ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
+import { INotificationService } from 'vs/platform/notification/common/notification';
+import { IProgress, Progress } from 'vs/platform/progress/common/progress';
+import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
+import { CodeActionFilter, CodeActionItem, CodeActionKind, CodeActionSet, CodeActionTrigger, CodeActionTriggerSource, filtersAction, mayIncludeActionsOfKind } from '../common/types';
 
 export const codeActionCommandId = 'editor.action.codeAction';
 export const refactorCommandId = 'editor.action.refactor';
@@ -27,36 +33,8 @@ export const sourceActionCommandId = 'editor.action.sourceAction';
 export const organizeImportsCommandId = 'editor.action.organizeImports';
 export const fixAllCommandId = 'editor.action.fixAll';
 
-export class CodeActionItem {
-
-	constructor(
-		readonly action: languages.CodeAction,
-		readonly provider: languages.CodeActionProvider | undefined,
-	) { }
-
-	async resolve(token: CancellationToken): Promise<this> {
-		if (this.provider?.resolveCodeAction && !this.action.edit) {
-			let action: languages.CodeAction | undefined | null;
-			try {
-				action = await this.provider.resolveCodeAction(this.action, token);
-			} catch (err) {
-				onUnexpectedExternalError(err);
-			}
-			if (action) {
-				this.action.edit = action.edit;
-			}
-		}
-		return this;
-	}
-}
-
-export interface CodeActionSet extends IDisposable {
-	readonly validActions: readonly CodeActionItem[];
-	readonly allActions: readonly CodeActionItem[];
-	readonly hasAutoFix: boolean;
-
-	readonly documentation: readonly languages.Command[];
-}
+export const acceptSelectedCodeActionCommand = 'acceptSelectedCodeAction';
+export const previewSelectedCodeActionCommand = 'previewSelectedCodeAction';
 
 class ManagedCodeActionSet extends Disposable implements CodeActionSet {
 
@@ -100,7 +78,6 @@ class ManagedCodeActionSet extends Disposable implements CodeActionSet {
 		return this.validActions.some(({ action: fix }) => !!fix.kind && CodeActionKind.QuickFix.contains(new CodeActionKind(fix.kind)) && !!fix.isPreferred);
 	}
 }
-
 
 const emptyCodeActionsResponse = { actions: [] as CodeActionItem[], documentation: undefined };
 
@@ -245,6 +222,81 @@ function getDocumentationFromProvider(
 		}
 	}
 	return undefined;
+}
+
+export enum ApplyCodeActionReason {
+	OnSave = 'onSave',
+	FromProblemsView = 'fromProblemsView',
+	FromCodeActions = 'fromCodeActions'
+}
+
+export async function applyCodeAction(
+	accessor: ServicesAccessor,
+	item: CodeActionItem,
+	codeActionReason: ApplyCodeActionReason,
+	options?: { preview?: boolean; editor?: ICodeEditor },
+): Promise<void> {
+	const bulkEditService = accessor.get(IBulkEditService);
+	const commandService = accessor.get(ICommandService);
+	const telemetryService = accessor.get(ITelemetryService);
+	const notificationService = accessor.get(INotificationService);
+
+	type ApplyCodeActionEvent = {
+		codeActionTitle: string;
+		codeActionKind: string | undefined;
+		codeActionIsPreferred: boolean;
+		reason: ApplyCodeActionReason;
+	};
+	type ApplyCodeEventClassification = {
+		codeActionTitle: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'The display label of the applied code action' };
+		codeActionKind: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'The kind (refactor, quickfix) of the applied code action' };
+		codeActionIsPreferred: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'Was the code action marked as being a preferred action?' };
+		reason: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'The kind of action used to trigger apply code action.' };
+		owner: 'mjbvz';
+		comment: 'Event used to gain insights into which code actions are being triggered';
+	};
+
+	telemetryService.publicLog2<ApplyCodeActionEvent, ApplyCodeEventClassification>('codeAction.applyCodeAction', {
+		codeActionTitle: item.action.title,
+		codeActionKind: item.action.kind,
+		codeActionIsPreferred: !!item.action.isPreferred,
+		reason: codeActionReason,
+	});
+
+	await item.resolve(CancellationToken.None);
+
+	if (item.action.edit) {
+		await bulkEditService.apply(item.action.edit, {
+			editor: options?.editor,
+			label: item.action.title,
+			quotableLabel: item.action.title,
+			code: 'undoredo.codeAction',
+			respectAutoSaveConfig: true,
+			showPreview: options?.preview,
+		});
+	}
+
+	if (item.action.command) {
+		try {
+			await commandService.executeCommand(item.action.command.id, ...(item.action.command.arguments || []));
+		} catch (err) {
+			const message = asMessage(err);
+			notificationService.error(
+				typeof message === 'string'
+					? message
+					: nls.localize('applyCodeActionFailed', "An unknown error occurred while applying the code action"));
+		}
+	}
+}
+
+function asMessage(err: any): string | undefined {
+	if (typeof err === 'string') {
+		return err;
+	} else if (err instanceof Error && typeof err.message === 'string') {
+		return err.message;
+	} else {
+		return undefined;
+	}
 }
 
 CommandsRegistry.registerCommand('_executeCodeActionProvider', async function (accessor, resource: URI, rangeOrSelection: Range | Selection, kind?: string, itemResolveCount?: number): Promise<ReadonlyArray<languages.CodeAction>> {
